Wrap JCLI reqwest errors in RestAPI error
